src/requirements.txt

pandas==1.5.3
numpy==1.22.3
regex==2022.10.31

airflow==2.9.1
google-cloud-bigquery==3.23.1

pydantic==1.9.2 
urllib3==1.26.14
putz==2022.1

requests==2.28.1
requests-oauthlib==1.3.1
